set XSRF header on userprofile request
moves getCylcHeaders to utils/url as a common utility to be applied to all requests
